PO-8936 - payment terms refactor
