+++ src/main/java/uk/gov/hmcts/opal/service/opal/OpalDefendantAccountPaymentTermsService.java
@@ -8,10 +8,16 @@ import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.opal.dto.AddPaymentCardRequestResponse;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
 import uk.gov.hmcts.opal.dto.RecordType;
+import uk.gov.hmcts.opal.dto.request.AddDefendantAccountPaymentTermsRequest;
 import uk.gov.hmcts.opal.entity.defendantaccount.DefendantAccountEntity;
 import uk.gov.hmcts.opal.entity.PaymentCardRequestEntity;
+import uk.gov.hmcts.opal.entity.enforcement.EnforcementEntity;
 import uk.gov.hmcts.opal.entity.paymentterms.PaymentTermsEntity;
 import uk.gov.hmcts.opal.controllers.advice.GlobalExceptionHandler.PaymentCardRequestAlreadyExistsException;
+import uk.gov.hmcts.opal.entity.result.ResultEntity;
+import uk.gov.hmcts.opal.mapper.request.PaymentTermsMapper;
+import uk.gov.hmcts.opal.repository.DefendantAccountRepository;
+import uk.gov.hmcts.opal.repository.EnforcementRepository;
 import uk.gov.hmcts.opal.service.iface.DefendantAccountPaymentTermsServiceInterface;
 import uk.gov.hmcts.opal.service.persistence.AmendmentRepositoryService;
 import uk.gov.hmcts.opal.service.persistence.DefendantAccountRepositoryService;
@@ -33,6 +39,22 @@ public class OpalDefendantAccountPaymentTermsService implements DefendantAccount
     private final AmendmentRepositoryService amendmentRepositoryService;
 
     private final PaymentCardRequestRepositoryService paymentCardRequestRepositoryService;
+
+    private final DocumentService documentService;
+
+    private final PaymentTermsService paymentTermsService;
+
+    private final AmendmentService amendmentService;
+
+    private final ReportEntryService reportEntryService;
+
+    private final PaymentTermsMapper paymentTermsMapper;
+
+    private final DefendantAccountRepository defendantAccountRepository;
+
+    private final EnforcementRepository enforcementRepository;
+
+    private final ResultService resultService;
 
     @Override
     @Transactional(readOnly = true)
@@ -131,4 +153,237 @@ public class OpalDefendantAccountPaymentTermsService implements DefendantAccount
             "ACCOUNT_ENQUIRY"
         );
     }
+
+    @Override
+    @Transactional
+    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
+    public GetDefendantAccountPaymentTermsResponse addPaymentTerms(Long defendantAccountId,
+        String businessUnitId,
+        String businessUnitUserId,
+        String postedByName,
+        String ifMatch,
+        AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest) {
+        log.debug(":addPaymentTerms (Opal): accountId={}, bu={}", defendantAccountId, businessUnitId);
+
+        // Look up the defendant account
+        DefendantAccountEntity defAccount = defendantAccountRepositoryService
+            .getDefendantAccountByIdForUpdate(defendantAccountId);
+
+        // Validate BU
+        if (defAccount.getBusinessUnit() == null
+            || defAccount.getBusinessUnit().getBusinessUnitId() == null
+            || !String.valueOf(defAccount.getBusinessUnit().getBusinessUnitId()).equals(businessUnitId)) {
+            throw new EntityNotFoundException("Defendant Account not found in business unit " + businessUnitId);
+        }
+
+        VersionUtils.verifyIfMatch(defAccount, ifMatch, defendantAccountId, "addPaymentTerms");
+
+        amendmentService.auditInitialiseStoredProc(defendantAccountId, RecordType.DEFENDANT_ACCOUNTS);
+
+        // Toggle any existing active payment term(s) for the defendant account to inactive
+        paymentTermsService.deactivateExistingActivePaymentTerms(defAccount.getDefendantAccountId());
+
+        // Map request -> Payment Terms Entity using MapStruct
+        PaymentTermsEntity paymentTermsEntity
+            = paymentTermsMapper.toEntity(addPaymentTermsRequest.getPaymentTerms());
+        paymentTermsEntity.setDefendantAccount(defAccount);
+        // If not included in the request, infer these from the token/user context
+        if (paymentTermsEntity.getPostedByUsername() == null) {
+            paymentTermsEntity.setPostedByUsername(businessUnitUserId);
+        }
+        if (paymentTermsEntity.getPostedBy() == null) {
+            paymentTermsEntity.setPostedBy(businessUnitUserId);
+        }
+        // Persist the new (active) PaymentTermsEntity
+        final PaymentTermsEntity savedPaymentTerms = paymentTermsService.addPaymentTerm(paymentTermsEntity);
+
+        // Update defendant account with any payment term related attributes
+        addPaymentTerm(defAccount, addPaymentTermsRequest);
+
+        // Clear last_enforcement on the defendant account, if applicable
+        clearLastEnforcementAction(defAccount);
+
+        defendantAccountRepository.save(defAccount);
+
+        // If requestPaymentCardFlag is true: create a PaymentCardRequest row (if not already present)
+        //  and update the defendant account PCR-related attributes (requested flag/date/by/byName).
+        if (Boolean.TRUE.equals(addPaymentTermsRequest.getRequestPaymentCard())) {
+            log.debug(":addPaymentTerms: Request Payment Card flag is TRUE for account {}",
+                defAccount.getDefendantAccountId());
+            addPaymentCard(defendantAccountId, businessUnitId, businessUnitUserId, ifMatch, postedByName);
+        }
+
+        // if generate_payment_terms_change_letter is true
+        //   Create a Document Instances record for a Payment Terms Change Letter.
+        if (Boolean.TRUE.equals(addPaymentTermsRequest.getGeneratePaymentTermsChangeLetter())) {
+            log.debug(":addPaymentTerms: Generate Payment Terms Change Letter flag is TRUE for account {}",
+                defAccount.getDefendantAccountId());
+            documentService.createDocumentInstance(defendantAccountId,
+                defAccount.getBusinessUnit().getBusinessUnitId());
+        }
+
+        // Create report entry for the Extension of Time to Pay report
+        reportEntryService.createExtendTtpReportEntry(defAccount.getDefendantAccountId(),
+            defAccount.getBusinessUnit().getBusinessUnitId());
+
+        log.debug(":addPaymentTerms: saved payment terms id={} for account {}",
+            savedPaymentTerms.getPaymentTermsId(), defAccount.getDefendantAccountId());
+
+        amendmentService.auditFinaliseStoredProc(
+            defAccount.getDefendantAccountId(),
+            RecordType.DEFENDANT_ACCOUNTS,
+            Short.parseShort(businessUnitId),
+            businessUnitUserId,
+            postedByName,
+            defAccount.getProsecutorCaseReference(),
+            "ACCOUNT_ENQUIRY"
+        );
+
+        return OpalDefendantAccountBuilders.buildPaymentTermsResponse(savedPaymentTerms);
+    }
+
+    @Transactional
+    public GetDefendantAccountPaymentTermsResponse addPaymentTermsPreservingLastEnforcement(Long defendantAccountId,
+        String businessUnitId,
+        String businessUnitUserId,
+        String postedByName,
+        String ifMatch,
+        AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest) {
+
+        log.debug(":addPaymentTermsPreservingLastEnforcement (Opal): accountId={}, bu={}",
+            defendantAccountId, businessUnitId);
+
+        DefendantAccountEntity defAccount = defendantAccountRepositoryService
+            .getDefendantAccountByIdForUpdate(defendantAccountId);
+
+        if (defAccount.getBusinessUnit() == null
+            || defAccount.getBusinessUnit().getBusinessUnitId() == null
+            || !String.valueOf(defAccount.getBusinessUnit().getBusinessUnitId()).equals(businessUnitId)) {
+            throw new EntityNotFoundException("Defendant Account not found in business unit " + businessUnitId);
+        }
+
+        VersionUtils.verifyIfMatch(defAccount, ifMatch, defendantAccountId, "addPaymentTerms");
+
+        amendmentService.auditInitialiseStoredProc(defendantAccountId, RecordType.DEFENDANT_ACCOUNTS);
+
+        paymentTermsService.deactivateExistingActivePaymentTerms(defAccount.getDefendantAccountId());
+
+        PaymentTermsEntity paymentTermsEntity = paymentTermsMapper.toEntity(addPaymentTermsRequest.getPaymentTerms());
+        paymentTermsEntity.setDefendantAccount(defAccount);
+        if (paymentTermsEntity.getPostedByUsername() == null) {
+            paymentTermsEntity.setPostedByUsername(businessUnitUserId);
+        }
+        if (paymentTermsEntity.getPostedBy() == null) {
+            paymentTermsEntity.setPostedBy(businessUnitUserId);
+        }
+
+        final PaymentTermsEntity savedPaymentTerms = paymentTermsService.addPaymentTerm(paymentTermsEntity);
+
+        addPaymentTerm(defAccount, addPaymentTermsRequest);
+        defendantAccountRepository.save(defAccount);
+
+        if (Boolean.TRUE.equals(addPaymentTermsRequest.getRequestPaymentCard())) {
+            log.debug(":addPaymentTermsPreservingLastEnforcement: Request Payment Card flag is TRUE for account {}",
+                defAccount.getDefendantAccountId());
+            addPaymentCard(defendantAccountId, businessUnitId, businessUnitUserId, ifMatch, postedByName);
+        }
+
+        if (Boolean.TRUE.equals(addPaymentTermsRequest.getGeneratePaymentTermsChangeLetter())) {
+            log.debug(":addPaymentTermsPreservingLastEnforcement: Generate Payment Terms Change Letter flag is TRUE "
+                    + "for account {}",
+                defAccount.getDefendantAccountId());
+            documentService.createDocumentInstance(defendantAccountId,
+                defAccount.getBusinessUnit().getBusinessUnitId());
+        }
+
+        reportEntryService.createExtendTtpReportEntry(savedPaymentTerms.getPaymentTermsId(),
+            defAccount.getBusinessUnit().getBusinessUnitId());
+
+        log.debug(":addPaymentTermsPreservingLastEnforcement: saved payment terms id={} for account {}",
+            savedPaymentTerms.getPaymentTermsId(), defAccount.getDefendantAccountId());
+
+        amendmentService.auditFinaliseStoredProc(
+            defAccount.getDefendantAccountId(),
+            RecordType.DEFENDANT_ACCOUNTS,
+            Short.parseShort(businessUnitId),
+            businessUnitUserId,
+            postedByName,
+            defAccount.getProsecutorCaseReference(),
+            "ACCOUNT_ENQUIRY"
+        );
+
+        return OpalDefendantAccountBuilders.buildPaymentTermsResponse(savedPaymentTerms);
+    }
+
+    /**
+     * This method adds a payment card request to a defendant account.
+     *
+     * <p>
+     * This method is separated from the public interface to allow for
+     * addition of a payment card in a chained operation.
+     * </p>
+     */
+    private AddPaymentCardRequestResponse addPaymentCard(
+        Long defendantAccountId,
+        String businessUnitId,
+        String businessUnitUserId,
+        String ifMatch,
+        String displayName) {
+
+        log.debug(":addPaymentCard (Opal): accountId={}, bu={}", defendantAccountId, businessUnitId);
+
+        DefendantAccountEntity account = loadAndValidateAccount(defendantAccountId, businessUnitId);
+        VersionUtils.verifyIfMatch(account, ifMatch, account.getDefendantAccountId(), "addPaymentCard");
+
+        ensureNoExistingPaymentCardRequest(defendantAccountId);
+
+        createPaymentCardRequest(defendantAccountId);
+
+        updateDefendantAccountWithPcr(account, businessUnitUserId, displayName);
+
+        // Minimal response
+        return new AddPaymentCardRequestResponse(defendantAccountId);
+    }
+
+
+    /**
+     * Add payment term related attributes to the defendant account.
+     */
+    private void addPaymentTerm(DefendantAccountEntity defAccount,
+        AddDefendantAccountPaymentTermsRequest paymentTermsRequest) {
+
+        defAccount.setSuspendedCommittalDate(paymentTermsRequest.getPaymentTerms().getDateDaysInDefaultImposed());
+    }
+
+    /**
+     * Along with adding a new latest payment_terms we will also clear the last enforcement action on the account.
+     * The only time it doesn't is when the last enforcement has result.extend_ttp_preserve_last_enf = TRUE
+     */
+    private void clearLastEnforcementAction(DefendantAccountEntity defAccount) {
+        // Retrieve most recent enforcement action for this account using the result
+        EnforcementEntity mostRecentEnforcement = fetchEnforcementMostRecent(defAccount);
+
+        if (mostRecentEnforcement != null) {
+            ResultEntity resultEntityLite = resultService.getResultById(mostRecentEnforcement.getResultId());
+
+            // If resultEntity exists and extend_ttp_preserve_last_enf is not TRUE, clear last_enforcement
+            if (!resultEntityLite.isExtendTtpPreserveLastEnf()) {
+                log.debug(":clearLastEnforcementAction: Clearing last_enforcement={} for account {}",
+                    defAccount.getLastEnforcement(), defAccount.getDefendantAccountId());
+                defAccount.setLastEnforcement(null);
+
+            } else {
+                log.debug(":clearLastEnforcementAction: Preserving last_enforcement={} for account {} as "
+                        + "extend_ttp_preserve_last_enf=TRUE",
+                    defAccount.getLastEnforcement(), defAccount.getDefendantAccountId());
+            }
+        }
+    }
+
+    private EnforcementEntity fetchEnforcementMostRecent(DefendantAccountEntity entity) {
+        return enforcementRepository.findFirstByDefendantAccountIdAndResultIdOrderByPostedDateDesc(
+            entity.getDefendantAccountId(), entity.getLastEnforcement()).orElse(null);
+    }
+
+
 }
