+++ src/main/java/uk/gov/hmcts/opal/service/opal/OpalDefendantAccountService.java
@@ -15,7 +15,6 @@ import jakarta.persistence.LockModeType;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.Clock;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
@@ -27,10 +26,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import uk.gov.hmcts.opal.controllers.advice.GlobalExceptionHandler.PaymentCardRequestAlreadyExistsException;
 import uk.gov.hmcts.opal.dto.AddDefendantAccountEnforcementRequest;
 import uk.gov.hmcts.opal.dto.AddEnforcementResponse;
-import uk.gov.hmcts.opal.dto.AddPaymentCardRequestResponse;
+
 import uk.gov.hmcts.opal.dto.DefendantAccountHeaderSummary;
 import uk.gov.hmcts.opal.dto.DefendantAccountSummaryDto;
 import uk.gov.hmcts.opal.dto.DefendantAccountSummaryDto.Checks;
@@ -39,14 +37,12 @@ import uk.gov.hmcts.opal.dto.DefendantAccountSummaryDto.WarnError;
 import uk.gov.hmcts.opal.dto.EnforcementStatus;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountConsolidatedAccountsResult;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountFixedPenaltyResponse;
-import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
 import uk.gov.hmcts.opal.dto.RecordType;
 import uk.gov.hmcts.opal.dto.UpdateDefendantAccountRequest;
 import uk.gov.hmcts.opal.dto.UpdateDefendantAccountResponse;
 import uk.gov.hmcts.opal.dto.common.EnforcementOverride;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryFilter;
 import uk.gov.hmcts.opal.dto.history.DefendantAccountHistoryResponse;
-import uk.gov.hmcts.opal.dto.request.AddDefendantAccountPaymentTermsRequest;
 import uk.gov.hmcts.opal.dto.response.DefendantAccountAtAGlanceResponse;
 import uk.gov.hmcts.opal.dto.search.AccountSearchDto;
 import uk.gov.hmcts.opal.dto.search.DefendantAccountSearchResultsDto;
@@ -54,14 +50,12 @@ import uk.gov.hmcts.opal.entity.EnforcerEntity;
 import uk.gov.hmcts.opal.entity.FixedPenaltyOffenceEntity;
 import uk.gov.hmcts.opal.entity.LocalJusticeAreaEntity;
 import uk.gov.hmcts.opal.entity.PartyEntity;
-import uk.gov.hmcts.opal.entity.PaymentCardRequestEntity;
 import uk.gov.hmcts.opal.entity.court.CourtEntity;
 import uk.gov.hmcts.opal.entity.debtordetail.DebtorDetailEntity;
 import uk.gov.hmcts.opal.entity.defendantaccount.DefendantAccountEntity;
 import uk.gov.hmcts.opal.entity.defendantaccount.DefendantAccountHeaderViewEntity;
 import uk.gov.hmcts.opal.entity.defendantaccount.DefendantAccountPartiesEntity;
 import uk.gov.hmcts.opal.entity.enforcement.EnforcementEntity;
-import uk.gov.hmcts.opal.entity.paymentterms.PaymentTermsEntity;
 import uk.gov.hmcts.opal.entity.result.ResultEntity;
 import uk.gov.hmcts.opal.entity.search.SearchConsolidatedEntity;
 import uk.gov.hmcts.opal.entity.search.SearchDefendantAccount;
@@ -74,24 +68,20 @@ import uk.gov.hmcts.opal.generated.model.UpdateDefendantAccountResponsePayload;
 import uk.gov.hmcts.opal.mapper.ConsolidatedAccountMapper;
 import uk.gov.hmcts.opal.mapper.DefendantAccountHeaderSummaryMapper;
 import uk.gov.hmcts.opal.mapper.common.EnforcerDefendantAccountMapper;
-import uk.gov.hmcts.opal.mapper.request.PaymentTermsMapper;
 import uk.gov.hmcts.opal.repository.ConsolidatedAccountRepository;
 import uk.gov.hmcts.opal.repository.CourtLiteRepository;
 import uk.gov.hmcts.opal.repository.DebtorDetailRepository;
-import uk.gov.hmcts.opal.repository.DefendantAccountPaymentTermsRepository;
 import uk.gov.hmcts.opal.repository.DefendantAccountRepository;
 import uk.gov.hmcts.opal.repository.EnforcementRepository;
 import uk.gov.hmcts.opal.repository.EnforcerRepository;
 import uk.gov.hmcts.opal.repository.FixedPenaltyOffenceRepository;
 import uk.gov.hmcts.opal.repository.LocalJusticeAreaRepository;
 import uk.gov.hmcts.opal.repository.NoteRepository;
-import uk.gov.hmcts.opal.repository.PaymentCardRequestRepository;
 import uk.gov.hmcts.opal.repository.ResultRepository;
 import uk.gov.hmcts.opal.repository.SearchDefendantBasicRepository;
 import uk.gov.hmcts.opal.repository.SearchDefendantConsolidatedRepository;
 import uk.gov.hmcts.opal.repository.jpa.SearchBasicEntitySpecs;
 import uk.gov.hmcts.opal.repository.jpa.SearchConsolidatedEntitySpecs;
-import uk.gov.hmcts.opal.service.UserStateService;
 import uk.gov.hmcts.opal.service.iface.DefendantAccountServiceInterface;
 import uk.gov.hmcts.opal.service.opal.history.defendant.DefendantAccountHistoryService;
 import uk.gov.hmcts.opal.service.persistence.DefendantAccountHeaderViewRepositoryService;
@@ -116,8 +106,6 @@ public class OpalDefendantAccountService implements DefendantAccountServiceInter
     private final SearchBasicEntitySpecs searchBasicEntitySpecs;
     private final SearchConsolidatedEntitySpecs searchConsolidatedEntitySpecs;
 
-    private final DefendantAccountPaymentTermsRepository defendantAccountPaymentTermsRepository;
-
     private final DefendantAccountSummaryViewRepositoryService defendantAccountSummaryViewRepositoryService;
     private final DefendantAccountHeaderViewRepositoryService defendantAccountHeaderViewRepositoryService;
 
@@ -138,21 +126,9 @@ public class OpalDefendantAccountService implements DefendantAccountServiceInter
 
     private final DebtorDetailRepository debtorDetailRepository;
 
-    private final PaymentCardRequestRepository paymentCardRequestRepository;
-
-    private final UserStateService userStateService;
-
     private final ResultRepository resultRepository;
 
     // Services
-    private final DocumentService documentService;
-
-    private final PaymentTermsService paymentTermsService;
-
-    private final ResultService resultService;
-
-    private final ReportEntryService reportEntryService;
-
     private final DefendantAccountHistoryService defendantAccountHistoryService;
 
     private final DefendantAccountRepositoryService defendantAccountRepositoryService;
@@ -161,7 +137,6 @@ public class OpalDefendantAccountService implements DefendantAccountServiceInter
     private final ConsolidatedAccountMapper consolidatedAccountMapper;
     private final DefendantAccountHeaderSummaryMapper defendantAccountHeaderSummaryMapper;
     private final EnforcerDefendantAccountMapper enforcerDefendantAccountMapper;
-    private final PaymentTermsMapper paymentTermsMapper;
 
     private final Clock clock;
 
@@ -290,22 +265,6 @@ public class OpalDefendantAccountService implements DefendantAccountServiceInter
             .aliases(OpalDefendantAccountBuilders.buildSearchAliases(account));
     }
 
-    //Deprecated - use OpalDefendantAccountPaymentTermsService
-    //TODO - Remove once OpalDefendantAccountPaymentTermsService is in use
-    @Override
-    @Transactional(readOnly = true)
-    public GetDefendantAccountPaymentTermsResponse getPaymentTerms(Long defendantAccountId) {
-        log.debug(":getPaymentTerms (Opal): criteria: {}", defendantAccountId);
-
-        PaymentTermsEntity entity = defendantAccountPaymentTermsRepository
-            .findTopByDefendantAccount_DefendantAccountIdOrderByPostedDateDescPaymentTermsIdDesc(
-                defendantAccountId)
-            .orElseThrow(() -> new EntityNotFoundException("Payment Terms not found for Defendant Account Id: "
-                + defendantAccountId));
-
-        return OpalDefendantAccountBuilders.buildPaymentTermsResponse(entity);
-    }
-
     //Deprecated - use OpalDefendantAccountFixedPenaltyService
     //TODO - Remove once OpalDefendantAccountFixedPenaltyService is in use
     @Override
@@ -401,36 +360,6 @@ public class OpalDefendantAccountService implements DefendantAccountServiceInter
             )
             .version(newVersion)
             .build();
-    }
-
-    /**
-     * This method adds a payment card request to a defendant account.
-     *
-     * <p>
-     * This method is separated from the public interface to allow for
-     * addition of a payment card in a chained operation.
-     * </p>
-     */
-    private AddPaymentCardRequestResponse addPaymentCard(
-        Long defendantAccountId,
-        String businessUnitId,
-        String businessUnitUserId,
-        String ifMatch,
-        String displayName) {
-
-        log.debug(":addPaymentCard (Opal): accountId={}, bu={}", defendantAccountId, businessUnitId);
-
-        DefendantAccountEntity account = loadAndValidateAccount(defendantAccountId, businessUnitId);
-        VersionUtils.verifyIfMatch(account, ifMatch, account.getDefendantAccountId(), "addPaymentCard");
-
-        ensureNoExistingPaymentCardRequest(defendantAccountId);
-
-        createPaymentCardRequest(defendantAccountId);
-
-        updateDefendantAccountWithPcr(account, businessUnitUserId, displayName);
-
-        // Minimal response
-        return new AddPaymentCardRequestResponse(defendantAccountId);
     }
 
     //Deprecated - use OpalDefendantAccountEnforcementService
@@ -550,299 +479,6 @@ public class OpalDefendantAccountService implements DefendantAccountServiceInter
         entity.setEnforcingCourt(court);
         log.debug(":applyEnforcementCourt: accountId={}, courtId={}",
             entity.getDefendantAccountId(), court.getCourtId());
-    }
-
-    //Deprecated - use DefendantAccountPaymentTermsService
-    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
-    @Override
-    @Transactional
-    public AddPaymentCardRequestResponse addPaymentCardRequest(Long defendantAccountId,
-        String businessUnitId,
-        String businessUnitUserId,
-        String postedByName,
-        String ifMatch) {
-
-        log.debug(":addPaymentCardRequest (Opal): accountId={}, bu={}", defendantAccountId, businessUnitId);
-
-        DefendantAccountEntity account = defendantAccountRepositoryService.findById(defendantAccountId);
-
-        amendmentService.auditInitialiseStoredProc(defendantAccountId, RecordType.DEFENDANT_ACCOUNTS);
-
-        AddPaymentCardRequestResponse paymentCardResponse = addPaymentCard(
-            defendantAccountId,
-            businessUnitId,
-            businessUnitUserId,
-            ifMatch,
-            postedByName
-        );
-
-        auditComplete(defendantAccountId, account, businessUnitUserId, postedByName);
-
-        return paymentCardResponse;
-    }
-
-    //Deprecated - use DefendantAccountPaymentTermsService
-    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
-    private DefendantAccountEntity loadAndValidateAccount(Long accountId, String buId) {
-        DefendantAccountEntity account = defendantAccountRepositoryService.findById(accountId);
-        validateBusinessUnitPresent(account, buId);
-        return account;
-    }
-
-    //Deprecated - use DefendantAccountPaymentTermsService
-    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
-    private void validateBusinessUnitPresent(DefendantAccountEntity account, String buId) {
-        if (account.getBusinessUnit() == null
-            || account.getBusinessUnit().getBusinessUnitId() == null
-            || !String.valueOf(account.getBusinessUnit().getBusinessUnitId()).equals(buId)) {
-            throw new EntityNotFoundException("Defendant Account not found in business unit " + buId);
-        }
-    }
-
-    //Deprecated - use DefendantAccountPaymentTermsService
-    private void ensureNoExistingPaymentCardRequest(Long accountId) {
-        if (paymentCardRequestRepository.existsByDefendantAccountId(accountId)) {
-            throw new PaymentCardRequestAlreadyExistsException(
-                "DefendantAccountEntity",
-                String.valueOf(accountId)
-            );
-        }
-    }
-
-    //Deprecated - use DefendantAccountPaymentTermsService
-    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
-    private void createPaymentCardRequest(Long accountId) {
-        PaymentCardRequestEntity pcr = PaymentCardRequestEntity.builder()
-            .defendantAccountId(accountId)
-            .build();
-        paymentCardRequestRepository.save(pcr);
-    }
-
-    //Deprecated - use DefendantAccountPaymentTermsService
-    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
-    private void updateDefendantAccountWithPcr(DefendantAccountEntity account,
-        String businessUnitUserId,
-        String displayName) {
-
-        account.setPaymentCardRequested(true);
-        account.setPaymentCardRequestedDate(LocalDate.now());
-        account.setPaymentCardRequestedBy(businessUnitUserId);
-        account.setPaymentCardRequestedByName(displayName);
-
-        defendantAccountRepository.save(account);
-    }
-
-    private void auditComplete(Long accountId,
-        DefendantAccountEntity account,
-        String businessUnitUserId,
-        String postedByName) {
-
-        Short buId = account.getBusinessUnit().getBusinessUnitId();
-
-        amendmentService.auditFinaliseStoredProc(
-            accountId,
-            RecordType.DEFENDANT_ACCOUNTS,
-            buId,
-            businessUnitUserId,
-            postedByName,
-            account.getProsecutorCaseReference(),
-            "ACCOUNT_ENQUIRY"
-        );
-    }
-
-    @Override
-    @Transactional
-    //TODO: Remove this method once OpalDefendantAccountPaymentTermsService is in use
-    public GetDefendantAccountPaymentTermsResponse addPaymentTerms(Long defendantAccountId,
-        String businessUnitId,
-        String businessUnitUserId,
-        String postedByName,
-        String ifMatch,
-        AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest) {
-        log.debug(":addPaymentTerms (Opal): accountId={}, bu={}", defendantAccountId, businessUnitId);
-
-        // Look up the defendant account
-        DefendantAccountEntity defAccount = defendantAccountRepositoryService
-            .getDefendantAccountByIdForUpdate(defendantAccountId);
-
-        // Validate BU
-        if (defAccount.getBusinessUnit() == null
-            || defAccount.getBusinessUnit().getBusinessUnitId() == null
-            || !String.valueOf(defAccount.getBusinessUnit().getBusinessUnitId()).equals(businessUnitId)) {
-            throw new EntityNotFoundException("Defendant Account not found in business unit " + businessUnitId);
-        }
-
-        VersionUtils.verifyIfMatch(defAccount, ifMatch, defendantAccountId, "addPaymentTerms");
-
-        amendmentService.auditInitialiseStoredProc(defendantAccountId, RecordType.DEFENDANT_ACCOUNTS);
-
-        // Toggle any existing active payment term(s) for the defendant account to inactive
-        paymentTermsService.deactivateExistingActivePaymentTerms(defAccount.getDefendantAccountId());
-
-        // Map request -> Payment Terms Entity using MapStruct
-        PaymentTermsEntity paymentTermsEntity
-            = paymentTermsMapper.toEntity(addPaymentTermsRequest.getPaymentTerms());
-        paymentTermsEntity.setDefendantAccount(defAccount);
-        // If not included in the request, infer these from the token/user context
-        if (paymentTermsEntity.getPostedByUsername() == null) {
-            paymentTermsEntity.setPostedByUsername(businessUnitUserId);
-        }
-        if (paymentTermsEntity.getPostedBy() == null) {
-            paymentTermsEntity.setPostedBy(businessUnitUserId);
-        }
-        // Persist the new (active) PaymentTermsEntity
-        final PaymentTermsEntity savedPaymentTerms = paymentTermsService.addPaymentTerm(paymentTermsEntity);
-
-        // Update defendant account with any payment term related attributes
-        addPaymentTerm(defAccount, addPaymentTermsRequest);
-
-        // Clear last_enforcement on the defendant account, if applicable
-        clearLastEnforcementAction(defAccount);
-
-        defendantAccountRepository.save(defAccount);
-
-        // If requestPaymentCardFlag is true: create a PaymentCardRequest row (if not already present)
-        //  and update the defendant account PCR-related attributes (requested flag/date/by/byName).
-        if (Boolean.TRUE.equals(addPaymentTermsRequest.getRequestPaymentCard())) {
-            log.debug(":addPaymentTerms: Request Payment Card flag is TRUE for account {}",
-                defAccount.getDefendantAccountId());
-            addPaymentCard(defendantAccountId, businessUnitId, businessUnitUserId, ifMatch, postedByName);
-        }
-
-        // if generate_payment_terms_change_letter is true
-        //   Create a Document Instances record for a Payment Terms Change Letter.
-        if (Boolean.TRUE.equals(addPaymentTermsRequest.getGeneratePaymentTermsChangeLetter())) {
-            log.debug(":addPaymentTerms: Generate Payment Terms Change Letter flag is TRUE for account {}",
-                defAccount.getDefendantAccountId());
-            documentService.createDocumentInstance(defendantAccountId,
-                defAccount.getBusinessUnit().getBusinessUnitId());
-        }
-
-        // Create report entry for the Extension of Time to Pay report
-        reportEntryService.createExtendTtpReportEntry(defAccount.getDefendantAccountId(),
-            defAccount.getBusinessUnit().getBusinessUnitId());
-
-        log.debug(":addPaymentTerms: saved payment terms id={} for account {}",
-            savedPaymentTerms.getPaymentTermsId(), defAccount.getDefendantAccountId());
-
-        amendmentService.auditFinaliseStoredProc(
-            defAccount.getDefendantAccountId(),
-            RecordType.DEFENDANT_ACCOUNTS,
-            Short.parseShort(businessUnitId),
-            businessUnitUserId,
-            postedByName,
-            defAccount.getProsecutorCaseReference(),
-            "ACCOUNT_ENQUIRY"
-        );
-
-        return OpalDefendantAccountBuilders.buildPaymentTermsResponse(savedPaymentTerms);
-    }
-
-    @Transactional
-    public GetDefendantAccountPaymentTermsResponse addPaymentTermsPreservingLastEnforcement(Long defendantAccountId,
-        String businessUnitId,
-        String businessUnitUserId,
-        String postedByName,
-        String ifMatch,
-        AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest) {
-
-        log.debug(":addPaymentTermsPreservingLastEnforcement (Opal): accountId={}, bu={}",
-            defendantAccountId, businessUnitId);
-
-        DefendantAccountEntity defAccount = defendantAccountRepositoryService
-            .getDefendantAccountByIdForUpdate(defendantAccountId);
-
-        if (defAccount.getBusinessUnit() == null
-            || defAccount.getBusinessUnit().getBusinessUnitId() == null
-            || !String.valueOf(defAccount.getBusinessUnit().getBusinessUnitId()).equals(businessUnitId)) {
-            throw new EntityNotFoundException("Defendant Account not found in business unit " + businessUnitId);
-        }
-
-        VersionUtils.verifyIfMatch(defAccount, ifMatch, defendantAccountId, "addPaymentTerms");
-
-        amendmentService.auditInitialiseStoredProc(defendantAccountId, RecordType.DEFENDANT_ACCOUNTS);
-
-        paymentTermsService.deactivateExistingActivePaymentTerms(defAccount.getDefendantAccountId());
-
-        PaymentTermsEntity paymentTermsEntity = paymentTermsMapper.toEntity(addPaymentTermsRequest.getPaymentTerms());
-        paymentTermsEntity.setDefendantAccount(defAccount);
-        if (paymentTermsEntity.getPostedByUsername() == null) {
-            paymentTermsEntity.setPostedByUsername(businessUnitUserId);
-        }
-        if (paymentTermsEntity.getPostedBy() == null) {
-            paymentTermsEntity.setPostedBy(businessUnitUserId);
-        }
-
-        final PaymentTermsEntity savedPaymentTerms = paymentTermsService.addPaymentTerm(paymentTermsEntity);
-
-        addPaymentTerm(defAccount, addPaymentTermsRequest);
-        defendantAccountRepository.save(defAccount);
-
-        if (Boolean.TRUE.equals(addPaymentTermsRequest.getRequestPaymentCard())) {
-            log.debug(":addPaymentTermsPreservingLastEnforcement: Request Payment Card flag is TRUE for account {}",
-                defAccount.getDefendantAccountId());
-            addPaymentCard(defendantAccountId, businessUnitId, businessUnitUserId, ifMatch, postedByName);
-        }
-
-        if (Boolean.TRUE.equals(addPaymentTermsRequest.getGeneratePaymentTermsChangeLetter())) {
-            log.debug(":addPaymentTermsPreservingLastEnforcement: Generate Payment Terms Change Letter flag is TRUE "
-                    + "for account {}",
-                defAccount.getDefendantAccountId());
-            documentService.createDocumentInstance(defendantAccountId,
-                defAccount.getBusinessUnit().getBusinessUnitId());
-        }
-
-        reportEntryService.createExtendTtpReportEntry(savedPaymentTerms.getPaymentTermsId(),
-            defAccount.getBusinessUnit().getBusinessUnitId());
-
-        log.debug(":addPaymentTermsPreservingLastEnforcement: saved payment terms id={} for account {}",
-            savedPaymentTerms.getPaymentTermsId(), defAccount.getDefendantAccountId());
-
-        amendmentService.auditFinaliseStoredProc(
-            defAccount.getDefendantAccountId(),
-            RecordType.DEFENDANT_ACCOUNTS,
-            Short.parseShort(businessUnitId),
-            businessUnitUserId,
-            postedByName,
-            defAccount.getProsecutorCaseReference(),
-            "ACCOUNT_ENQUIRY"
-        );
-
-        return OpalDefendantAccountBuilders.buildPaymentTermsResponse(savedPaymentTerms);
-    }
-
-    /**
-     * Along with adding a new latest payment_terms we will also clear the last enforcement action on the account.
-     * The only time it doesn't is when the last enforcement has result.extend_ttp_preserve_last_enf = TRUE
-     */
-    private void clearLastEnforcementAction(DefendantAccountEntity defAccount) {
-        // Retrieve most recent enforcement action for this account using the result
-        EnforcementEntity mostRecentEnforcement = fetchEnforcementMostRecent(defAccount);
-
-        if (mostRecentEnforcement != null) {
-            ResultEntity resultEntityLite = resultService.getResultById(mostRecentEnforcement.getResultId());
-
-            // If resultEntity exists and extend_ttp_preserve_last_enf is not TRUE, clear last_enforcement
-            if (!resultEntityLite.isExtendTtpPreserveLastEnf()) {
-                log.debug(":clearLastEnforcementAction: Clearing last_enforcement={} for account {}",
-                    defAccount.getLastEnforcement(), defAccount.getDefendantAccountId());
-                defAccount.setLastEnforcement(null);
-
-            } else {
-                log.debug(":clearLastEnforcementAction: Preserving last_enforcement={} for account {} as "
-                        + "extend_ttp_preserve_last_enf=TRUE",
-                    defAccount.getLastEnforcement(), defAccount.getDefendantAccountId());
-            }
-        }
-    }
-
-    /**
-     * Add payment term related attributes to the defendant account.
-     */
-    private void addPaymentTerm(DefendantAccountEntity defAccount,
-        AddDefendantAccountPaymentTermsRequest paymentTermsRequest) {
-
-        defAccount.setSuspendedCommittalDate(paymentTermsRequest.getPaymentTerms().getDateDaysInDefaultImposed());
     }
 
     private boolean hasNonZeroBalance(DefendantAccountSummaryDto dto) {
