+++ src/test/java/uk/gov/hmcts/opal/service/legacy/LegacyDefAccServicePaymentTermsTest.java
@@ -3,6 +3,9 @@ package uk.gov.hmcts.opal.service.legacy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.when;
 
 import java.math.BigInteger;
@@ -14,6 +17,7 @@ import org.mockito.Mockito;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import uk.gov.hmcts.opal.common.legacy.service.GatewayService;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
 import uk.gov.hmcts.opal.dto.common.InstalmentPeriod;
 import uk.gov.hmcts.opal.dto.common.PaymentTermsType;
@@ -51,7 +55,7 @@ class LegacyDefAccServicePaymentTermsTest extends AbstractLegacyDefAccServiceTes
         when(restClient.responseSpec.toEntity(String.class))
             .thenReturn(new ResponseEntity<>(responseBody.toXml(), HttpStatus.OK));
 
-        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountService.getPaymentTerms(99L);
+        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountPaymentTermsService.getPaymentTerms(99L);
 
         assertNotNull(out);
         assertEquals(BigInteger.valueOf(2), out.getVersion());
@@ -84,7 +88,7 @@ class LegacyDefAccServicePaymentTermsTest extends AbstractLegacyDefAccServiceTes
         when(restClient.responseSpec.toEntity(String.class))
             .thenReturn(new ResponseEntity<>(responseBody.toXml(), HttpStatus.SERVICE_UNAVAILABLE));
 
-        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountService.getPaymentTerms(1L);
+        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountPaymentTermsService.getPaymentTerms(1L);
 
         assertNotNull(out);
         assertEquals(BigInteger.valueOf(3L), out.getVersion());
@@ -104,7 +108,7 @@ class LegacyDefAccServicePaymentTermsTest extends AbstractLegacyDefAccServiceTes
         when(restClient.responseSpec.toEntity(String.class))
             .thenReturn(new ResponseEntity<>("<error/>", HttpStatus.INTERNAL_SERVER_ERROR));
 
-        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountService.getPaymentTerms(42L);
+        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountPaymentTermsService.getPaymentTerms(42L);
 
         assertNull(out);
     }
@@ -118,7 +122,7 @@ class LegacyDefAccServicePaymentTermsTest extends AbstractLegacyDefAccServiceTes
         when(restClient.responseSpec.toEntity(String.class))
             .thenReturn(new ResponseEntity<>("<response/>", HttpStatus.OK));
 
-        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountService.getPaymentTerms(3L);
+        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountPaymentTermsService.getPaymentTerms(3L);
 
         assertNotNull(out);
         assertEquals(BigInteger.valueOf(1L), out.getVersion());
@@ -155,12 +159,34 @@ class LegacyDefAccServicePaymentTermsTest extends AbstractLegacyDefAccServiceTes
         when(restClient.responseSpec.toEntity(String.class))
             .thenReturn(new ResponseEntity<>(responseBody.toXml(), HttpStatus.OK));
 
-        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountService.getPaymentTerms(4L);
+        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountPaymentTermsService.getPaymentTerms(4L);
 
         assertNotNull(out);
         assertNotNull(out.getPaymentTerms());
         assertNull(out.getPaymentTerms().getPaymentTermsType());
         assertNull(out.getPaymentTerms().getInstalmentPeriod());
         assertNull(out.getPaymentTerms().getPostedDetails());
+    }
+
+    @Test
+    void legacyPaymentTerms_nonNullEnums_areConverted() {
+        LegacyGetDefendantAccountPaymentTermsResponse legacy = LegacyGetDefendantAccountPaymentTermsResponse.builder()
+            .version(1L)
+            .paymentTerms(
+                LegacyPaymentTerms.builder()
+                    .paymentTermsType(new LegacyPaymentTermsType(LegacyPaymentTermsType.PaymentTermsTypeCode.B))
+                    .instalmentPeriod(new LegacyInstalmentPeriod(LegacyInstalmentPeriod.InstalmentPeriodCode.W))
+                    .build()
+            )
+            .build();
+
+        doReturn(new GatewayService.Response<>(HttpStatus.OK, legacy, null, null))
+            .when(gatewayService).postToGateway(eq(LegacyDefendantAccountService.GET_PAYMENT_TERMS),
+                eq(LegacyGetDefendantAccountPaymentTermsResponse.class), any(), any());
+
+        GetDefendantAccountPaymentTermsResponse out = legacyDefendantAccountPaymentTermsService.getPaymentTerms(123L);
+
+        assertNotNull(out.getPaymentTerms().getPaymentTermsType());
+        assertNotNull(out.getPaymentTerms().getInstalmentPeriod());
     }
 }
